classifier/requirements.txt
tensorflow-lite-runtime==2.14.0
pillow==10.1.0
numpy==1.24.3
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
watchdog==3.0.0
python-dotenv==1.0.0
